Export field auth_token to packet
This is required for packet-ccm. packet-ccm needs secret to be
configured before it is applied. In that secret 2 values are required,
one is packet project_id and other is api_key. We were already getting
project_id and auth_token from user.

github.com/packethost/packet-ccm#get-packet-project-id-and-api-token
for more information.

Signed-off-by: knrt10 <kautilya@kinvolk.io>